move AIChat component back to widgets
