+++ app/client/packages/design-system/widgets/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./components/AIChat";
 export * from "./components/Icon";
 export * from "./components/Button";
 export * from "./components/IconButton";
