+++ app/client/src/modules/ui-builder/ui/wds/WDSAIChatWidget/AIChat/src/UserAvatar/styles.module.css
